requirements.txt
numpy
tensorflow-gpu==1.12